+++ DamPizzaApp/src/main/java/com/dampizza/App.java
@@ -12,8 +12,8 @@ import com.dampizza.views.login.RecoverView;
 import com.dampizza.views.login.SignupView;
 import com.dampizza.views.order.CartView;
 import com.dampizza.views.user.customer.CustomerView;
-import com.dampizza.views.user.manager.dealer.DealerView;
-import com.dampizza.views.user.manager.dealer.DealerOrderView;
+import com.dampizza.views.user.manager.DealerView;
+import com.dampizza.views.user.manager.DealerOrderView;
 import com.dampizza.views.user.customer.HistoryView;
 import com.dampizza.views.user.manager.ManagerView;
 import com.dampizza.views.user.manager.ManagerOrderView;
